Renamed 'testfunctions' to 'problem'
